move searchBukus func to utils.ts
