+++ utils/index.ts
@@ -36,6 +36,41 @@ export async function getBukus(typeId?: Kategori["id"]) {
   return data
 }
 
+interface BukuSearchArgs {
+  searchTerm?: string
+  category?: Kategori["id"] | null
+}
+
+/**
+ * search bukus.
+ * Both searchTerm and category is optional so when both isn't needed you need
+ * to pass an empty obj
+ */
+export async function searchBukus({ searchTerm, category }: BukuSearchArgs) {
+  const supabase = useSupabaseClient<Database>()
+
+  let query = supabase
+    .from("buku")
+    .select(`no_isbn, judul, penulis, penerbit, tahun_terbit, kategori_buku(kategori)`)
+
+  if (searchTerm) {
+    query = query.textSearch("judul", searchTerm, { type: "websearch" })
+  }
+
+  if (category) {
+    query = query.eq("kategori_id", category)
+  }
+
+  try {
+    const { data, error } = await query.limit(100)
+    if (error) throw error
+    return data
+  } catch (error) {
+    console.trace(error as PostgrestError)
+    return []
+  }
+}
+
 export async function getBukuImage(isbn: Buku["no_isbn"]) {
   const config = useRuntimeConfig()
   const supabase = useSupabaseClient<Database>()
